Moved WriteRDGPartHeader to tsuba
