+++ libtsuba/include/katana/tsuba.h
@@ -143,6 +143,26 @@ using EntityTypeIDToSetOfEntityTypeIDsStorageMap =
 /// can be increased if required
 constexpr size_t kMaxNumTopologies = 64;
 
+/// RDGPropInfo details the property name and its respective path in the RDG dir.
+struct RDGPropInfo {
+  std::string property_name;
+  std::string property_path;
+};
+
+// N.B. This is a temporary interface used to write RDG part header given some
+// amount of information regarding properties, type manager, etc. The primary consumer
+// of this interface is the out-of-core import path, which currently writes out property
+// and type information on its own.
+// TODO (vkarthik): get rid of this interface and have a proper unified one
+KATANA_EXPORT katana::Result<void> WriteRDGPartHeader(
+    std::vector<katana::RDGPropInfo> node_properties,
+    std::vector<katana::RDGPropInfo> edge_properties,
+    katana::EntityTypeManager& node_entity_type_manager,
+    katana::EntityTypeManager& edge_entity_type_manager,
+    const std::string& node_entity_type_id_array_path,
+    const std::string& edge_entity_type_id_array_path,
+    const std::string& topology_path, const std::string& rdg_dir);
+
 }  // namespace katana
 
 #endif
